@@ -9,6 +9,7 @@ import 'package:turbo/features/measuring/util/catmull_rom_spline.dart';
 import 'package:turbo/features/measuring/widgets/measuring_controls.dart';
 import 'package:turbo/features/measuring/widgets/measuring_line.dart';
 import 'package:turbo/features/measuring/widgets/measuring_markers.dart';
+import 'package:turbo/features/settings/api.dart';
 import 'package:turbo/widgets/map/controller/map_utility.dart';
 import 'package:turbo/widgets/map/controls/bottom_controls.dart';
 import 'package:turbo/widgets/map/controls/default_map_controls.dart';
@@ -17,12 +18,11 @@ import 'package:turbo/widgets/map/controls/map_controls.dart';
 
 import '../../map_view/widgets/map_base.dart';
 import '../../tile_providers/api.dart';
+import '../data/freehand_gesture_handler.dart';
+import '../util/edge_pan_handler.dart';
 
-final measuringStateProvider = StateNotifierProvider.autoDispose
-    .family<MeasuringStateNotifier, MeasuringState, LatLng>(
-      (ref, startPoint) {
-    return MeasuringStateNotifier(startPoint: startPoint);
-  },
+final measuringStateProvider = NotifierProvider.family<MeasuringStateNotifier, MeasuringState, LatLng>(
+  (arg) => MeasuringStateNotifier()..initialStartPoint = arg,
 );
 
 class MeasuringMapPage extends ConsumerStatefulWidget {
@@ -43,24 +43,37 @@ class MeasuringMapPage extends ConsumerStatefulWidget {
 
 class _MeasuringMapPageState extends ConsumerState<MeasuringMapPage>
     with TickerProviderStateMixin {
-  late MapController _mapController;
-  bool _isPointerDown = false;
-  Offset? _lastPointerScreenPos;
+  late final MapController _mapController;
+  late final EdgePanHandler _edgePanHandler;
+  late final FreehandGestureHandler _gestureHandler;
 
   @override
   void initState() {
     super.initState();
     _mapController = MapController();
+    
+    _edgePanHandler = EdgePanHandler(mapController: _mapController);
+    
+    _gestureHandler = FreehandGestureHandler(
+      notifier: ref.read(measuringStateProvider(widget.startPoint).notifier),
+      getSensitivity: () => ref.read(settingsProvider).value?.drawSensitivity ?? 15.0,
+      onEdgePanStop: () => _edgePanHandler.stop(),
+      onMove: (pos) => _edgePanHandler.handlePointerMove(pos, MediaQuery.of(context).size),
+    );
+  }
+
+  @override
+  void dispose() {
+    _edgePanHandler.dispose();
+    super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
     final layers = ref.watch(tileRegistryProvider.notifier).getActiveLayers();
     final registry = ref.watch(tileRegistryProvider);
-
-    final measuringNotifier =
-    ref.watch(measuringStateProvider(widget.startPoint).notifier);
     final measuringState = ref.watch(measuringStateProvider(widget.startPoint));
+    final measuringNotifier = ref.watch(measuringStateProvider(widget.startPoint).notifier);
 
     final width = MediaQuery.of(context).size.width;
     final isMobile = width < 600;
@@ -87,8 +100,6 @@ class _MeasuringMapPageState extends ConsumerState<MeasuringMapPage>
         initialCenter: widget.initialPosition,
         initialZoom: widget.zoom,
         mapController: _mapController,
-        // The onMapReady callback is removed to prevent test instability.
-        // The map already initializes at the correct location.
         mapLayers: [
           ...layers,
           ...registry.activeGlobalIds.map((id) => RichAttributionWidget(
@@ -102,42 +113,45 @@ class _MeasuringMapPageState extends ConsumerState<MeasuringMapPage>
           MeasureMarkers(points: pointsToRenderForMarkers)
         ],
         overlayWidgets: [
-          MapControls(controls: controls),
-          const Positioned(top: 16, left: 16, child: GoBackButton()),
+          MapControls(controls: controls, top: 16),
+          const Positioned(
+            top: 16,
+            left: 16,
+            child: GoBackButton(),
+          ),
           BottomControls(
             controls: MeasuringControls(
               distance: measuringState.totalDistance,
-              onReset: measuringNotifier.reset,
+              onReset: () {
+                _gestureHandler.reset();
+                measuringNotifier.reset();
+              },
               onUndo: measuringNotifier.undoLastPoint,
               onFinish: () => Navigator.of(context).pop(),
               onToggleSmoothing: measuringNotifier.toggleSmoothing,
               onToggleDrawing: measuringNotifier.toggleDrawing,
               onToggleIntermediatePoints:
               measuringNotifier.toggleIntermediatePoints,
-              onSensitivityChanged: measuringNotifier.setDrawSensitivity,
               canUndo: measuringState.points.length > 1,
               canReset: measuringState.points.length > 1,
               isSmoothing: measuringState.isSmoothing,
               isDrawing: measuringState.isDrawing,
               showIntermediatePoints: measuringState.showIntermediatePoints,
-              drawSensitivity: measuringState.drawSensitivity,
-
             ),
           ),
         ],
         onTap: measuringState.isDrawing
             ? null
             : (tapPosition, point) => _handleMapTap(point),
-        onPointerDown: measuringState.isDrawing ? _onPointerDown : null,
-        onPointerMove: measuringState.isDrawing ? _onPointerMove : null,
-        onPointerUp: measuringState.isDrawing ? _onPointerUp : null,
+        onPointerDown: measuringState.isDrawing ? _gestureHandler.handlePointerDown : null,
+        onPointerMove: measuringState.isDrawing ? _gestureHandler.handlePointerMove : null,
+        onPointerUp: measuringState.isDrawing ? _gestureHandler.handlePointerUp : null,
         interactionOptions: InteractionOptions(
           flags: measuringState.isDrawing
-              ? InteractiveFlag.none
-              : InteractiveFlag.all,
+              ? (InteractiveFlag.all & ~InteractiveFlag.drag & ~InteractiveFlag.rotate)
+              : (InteractiveFlag.all & ~InteractiveFlag.rotate),
           pinchZoomThreshold: 0.2,
           pinchMoveThreshold: 40,
-          rotationThreshold: 5.0,
         ),
       ),
     );
@@ -148,35 +162,5 @@ class _MeasuringMapPageState extends ConsumerState<MeasuringMapPage>
         .read(measuringStateProvider(widget.startPoint).notifier)
         .addPoint(point);
     animatedMapMove(point, _mapController.camera.zoom, _mapController, this);
-  }
-
-  void _onPointerDown(PointerDownEvent event, LatLng point) {
-    setState(() {
-      _isPointerDown = true;
-      _lastPointerScreenPos = event.localPosition;
-    });
-    ref.read(measuringStateProvider(widget.startPoint).notifier).addPoint(point);
-  }
-
-  void _onPointerMove(PointerMoveEvent event, LatLng point) {
-    if (!_isPointerDown || _lastPointerScreenPos == null) return;
-    final sensitivity =
-        ref.read(measuringStateProvider(widget.startPoint)).drawSensitivity;
-    final distance = (event.localPosition - _lastPointerScreenPos!).distance;
-    if (distance > sensitivity) {
-      ref
-          .read(measuringStateProvider(widget.startPoint).notifier)
-          .addPoint(point);
-      setState(() {
-        _lastPointerScreenPos = event.localPosition;
-      });
-    }
-  }
-
-  void _onPointerUp(PointerUpEvent event, LatLng point) {
-    setState(() {
-      _isPointerDown = false;
-      _lastPointerScreenPos = null;
-    });
   }
 }
