@@ -7,50 +7,67 @@ import 'package:geolocator/geolocator.dart';
 import 'package:latlong2/latlong.dart';
 
 final locationStateProvider =
-AutoDisposeAsyncNotifierProvider<LocationState, LatLng?>(
+AsyncNotifierProvider.autoDispose<LocationState, LatLng?>(
   LocationState.new,
 );
 
-class LocationState extends AutoDisposeAsyncNotifier<LatLng?> {
+class LocationState extends AsyncNotifier<LatLng?> {
   StreamSubscription<Position>? _positionSubscription;
 
   @override
   Future<LatLng?> build() async {
-    final completer = Completer<LatLng?>();
-
     ref.onDispose(() {
       _positionSubscription?.cancel();
     });
 
-    _setupLocationListener(completer);
-
-    return completer.future;
+    // Start listening but don't wait for the first value in build if it might hang.
+    // However, build() must return the initial value or wait for it.
+    // We'll use a timeout to prevent infinite hangs on problematic emulators.
+    try {
+      return await _setupLocationListener().timeout(
+        const Duration(seconds: 5),
+        onTimeout: () {
+          debugPrint("Location initialization timed out.");
+          return null;
+        },
+      );
+    } catch (e) {
+      debugPrint("Error in LocationState.build: $e");
+      return null;
+    }
   }
 
-  Future<void> _setupLocationListener(Completer<LatLng?> completer) async {
-    try {
-      if (!kIsWeb && Platform.isLinux) {
-        if (!completer.isCompleted) completer.complete(null);
-        return;
-      }
+  Future<LatLng?> _setupLocationListener() async {
+    if (!kIsWeb && Platform.isLinux) {
+      return null;
+    }
 
+    try {
       final serviceEnabled = await Geolocator.isLocationServiceEnabled();
       if (!serviceEnabled) {
-        throw Exception('location_services_disabled');
+        return null;
       }
 
       var permission = await Geolocator.checkPermission();
       if (permission == LocationPermission.denied) {
         permission = await Geolocator.requestPermission();
         if (permission == LocationPermission.denied) {
-          throw Exception('location_permissions_denied');
+          return null;
         }
       }
 
       if (permission == LocationPermission.deniedForever) {
-        throw Exception('location_permissions_denied_forever');
+        return null;
       }
 
+      // Get initial position
+      final position = await Geolocator.getCurrentPosition(
+        locationSettings: const LocationSettings(
+          accuracy: LocationAccuracy.high,
+        ),
+      );
+
+      // Start stream for updates
       _positionSubscription = Geolocator.getPositionStream(
         locationSettings: const LocationSettings(
           accuracy: LocationAccuracy.high,
@@ -58,24 +75,17 @@ class LocationState extends AutoDisposeAsyncNotifier<LatLng?> {
         ),
       ).listen(
             (Position position) {
-          final newPosition = LatLng(position.latitude, position.longitude);
-          if (!completer.isCompleted) {
-            completer.complete(newPosition);
-          }
-          state = AsyncData(newPosition);
+          state = AsyncData(LatLng(position.latitude, position.longitude));
         },
         onError: (error, stackTrace) {
-          if (!completer.isCompleted) {
-            completer.completeError(error, stackTrace);
-          }
-          state = AsyncError(error, stackTrace);
+          debugPrint("Location stream error: $error");
         },
       );
-    } catch (e, st) {
-      if (!completer.isCompleted) {
-        completer.completeError(e, st);
-      }
-      state = AsyncError(e, st);
+
+      return LatLng(position.latitude, position.longitude);
+    } catch (e) {
+      debugPrint("Location setup failed: $e");
+      return null;
     }
   }
 
@@ -86,11 +96,11 @@ class LocationState extends AutoDisposeAsyncNotifier<LatLng?> {
 }
 
 final locationSettingsProvProvider =
-AutoDisposeNotifierProvider<LocationSettingsProv, LocationAccuracy>(
+NotifierProvider.autoDispose<LocationSettingsProv, LocationAccuracy>(
   LocationSettingsProv.new,
 );
 
-class LocationSettingsProv extends AutoDisposeNotifier<LocationAccuracy> {
+class LocationSettingsProv extends Notifier<LocationAccuracy> {
   @override
   LocationAccuracy build() => LocationAccuracy.high;
 
