@@ -2,6 +2,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:turbo/data/auth/auth_providers.dart';
+import 'package:turbo/utils.dart';
 
 import 'login_screen.dart';
 import 'register_view_desktop.dart';
@@ -46,8 +47,8 @@ class _RegisterScreenState extends ConsumerState<RegisterScreen> {
   final _emailController = TextEditingController();
   final _passwordController = TextEditingController();
 
-  final _isLoading = StateProvider<bool>((ref) => false);
-  final _isGoogleLoading = StateProvider<bool>((ref) => false);
+  final _isLoading = NotifierProvider<LoadingNotifier, bool>(LoadingNotifier.new);
+  final _isGoogleLoading = NotifierProvider<LoadingNotifier, bool>(LoadingNotifier.new);
 
   @override
   void initState() {
@@ -80,7 +81,7 @@ class _RegisterScreenState extends ConsumerState<RegisterScreen> {
   Future<void> _register() async {
     FocusScope.of(context).unfocus();
     if (_formKey.currentState?.validate() ?? false) {
-      ref.read(_isLoading.notifier).state = true;
+      ref.read(_isLoading.notifier).set(true);
       try {
         await ref.read(authStateProvider.notifier).register(
           _emailController.text.trim(),
@@ -88,7 +89,7 @@ class _RegisterScreenState extends ConsumerState<RegisterScreen> {
         );
       } finally {
         if (mounted) {
-          ref.read(_isLoading.notifier).state = false;
+          ref.read(_isLoading.notifier).set(false);
         }
       }
     }
