@@ -2,10 +2,12 @@ import 'package:flutter/material.dart';
 
 class MapControls extends StatelessWidget {
   final List<Widget> controls;
+  final double? top;
 
   const MapControls({
     super.key,
     required this.controls,
+    this.top,
   });
 
   @override
@@ -13,15 +15,16 @@ class MapControls extends StatelessWidget {
     final width = MediaQuery.of(context).size.width;
     final isMobile = width < 600;
 
-    if(isMobile){
+    if (isMobile) {
       return Positioned(
-        top: 100,
+        top: top ?? 72,
         right: 16,
         child: Column(
+          mainAxisSize: MainAxisSize.min,
           children: controls,
         ),
       );
-    }else{
+    } else {
       return Positioned(
         bottom: 80,
         right: 16,
